+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/CreateObjectOptions.java
@@ -17,7 +17,8 @@
 
 package com.google.cloud.hadoop.gcsio;
 
-import com.google.common.base.Preconditions;
+import static com.google.common.base.Preconditions.checkArgument;
+
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
 
@@ -25,13 +26,14 @@ import java.util.Map;
  * Options for creating objects in GCS.
  */
 public class CreateObjectOptions {
-  public static final ImmutableMap<String, byte[]> EMPTY_METADATA =
-      ImmutableMap.<String, byte[]>of();
+  public static final ImmutableMap<String, byte[]> EMPTY_METADATA = ImmutableMap.of();
   public static final String DEFAULT_CONTENT_TYPE = "application/octet-stream";
+  public static final String DEFAULT_CONTENT_ENCODING = null;
   public static final CreateObjectOptions DEFAULT = new CreateObjectOptions(true);
 
   private final boolean overwriteExisting;
   private final String contentType;
+  private final String contentEncoding;
   private final Map<String, byte[]> metadata;
   private final boolean requireMetadataMatchForEmptyObjects;
 
@@ -43,7 +45,7 @@ public class CreateObjectOptions {
    * @param overwriteExisting True to overwrite any existing objects with the same name.
    */
   public CreateObjectOptions(boolean overwriteExisting) {
-    this(overwriteExisting, DEFAULT_CONTENT_TYPE, EMPTY_METADATA, false);
+    this(overwriteExisting, DEFAULT_CONTENT_TYPE, DEFAULT_CONTENT_ENCODING, EMPTY_METADATA, false);
   }
 
   /**
@@ -56,7 +58,7 @@ public class CreateObjectOptions {
    * @param metadata A dictionary of metadata to apply to created objects.
    */
   public CreateObjectOptions(boolean overwriteExisting, Map<String, byte[]> metadata) {
-    this(overwriteExisting, DEFAULT_CONTENT_TYPE, metadata, true);
+    this(overwriteExisting, DEFAULT_CONTENT_TYPE, metadata);
   }
 
   /**
@@ -70,7 +72,7 @@ public class CreateObjectOptions {
    */
   public CreateObjectOptions(
       boolean overwriteExisting, String contentType, Map<String, byte[]> metadata) {
-    this(overwriteExisting, contentType, metadata, true);
+    this(overwriteExisting, contentType, DEFAULT_CONTENT_ENCODING, metadata, true);
   }
 
   /**
@@ -78,22 +80,30 @@ public class CreateObjectOptions {
    *
    * @param overwriteExisting True to overwrite any existing objects with the same name
    * @param contentType content-type for the created file
+   * @param contentEncoding content-encoding for the created file
    * @param metadata A dictionary of metadata to apply to created objects
-   * @param requireMetadataMatchForEmptyObjects if true, when creating an empty object and
-   *     certain types of errors occur, any existing object is checked for an exact metadata
-   *     match to the metadata in this CreateObjectOptions before accepting the creation as
-   *     successful. If false, then on error for creating empty objects, as long as an
-   *     appropriate empty object already exists, even if it holds different metadata than
-   *     provided in this CreateObjectOptions instance, it may be considered created
-   *     successfully.
+   * @param requireMetadataMatchForEmptyObjects if true, when creating an empty object and certain
+   *     types of errors occur, any existing object is checked for an exact metadata match to the
+   *     metadata in this CreateObjectOptions before accepting the creation as successful. If false,
+   *     then on error for creating empty objects, as long as an appropriate empty object already
+   *     exists, even if it holds different metadata than provided in this CreateObjectOptions
+   *     instance, it may be considered created successfully.
    */
   public CreateObjectOptions(
-      boolean overwriteExisting, String contentType, Map<String, byte[]> metadata,
+      boolean overwriteExisting,
+      String contentType,
+      String contentEncoding,
+      Map<String, byte[]> metadata,
       boolean requireMetadataMatchForEmptyObjects) {
-    Preconditions.checkArgument(!metadata.containsKey("Content-Type"),
-        "The Content-Type metadata must be provided explicitly via the 'contentType' parameter");
+    checkArgument(
+        !metadata.containsKey("Content-Type"),
+        "The Content-Type must be provided explicitly via the 'contentType' parameter");
+    checkArgument(
+        !metadata.containsKey("Content-Encoding"),
+        "The Content-Encoding must be provided explicitly via the 'contentEncoding' parameter");
     this.overwriteExisting = overwriteExisting;
     this.contentType = contentType;
+    this.contentEncoding = contentEncoding;
     this.metadata = metadata;
     this.requireMetadataMatchForEmptyObjects = requireMetadataMatchForEmptyObjects;
   }
@@ -110,6 +120,11 @@ public class CreateObjectOptions {
    */
   public String getContentType() {
     return contentType;
+  }
+
+  /** Content type to set when creating a file. */
+  public String getContentEncoding() {
+    return contentEncoding;
   }
 
   /**