+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageImpl.java
@@ -376,10 +376,12 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
             clientRequestHelper,
             resourceId.getBucketName(),
             resourceId.getObjectName(),
+            options.getContentType(),
+            options.getContentEncoding(),
+            /* kmsKeyName= */ null,
             storageOptions.getWriteChannelOptions(),
             writeConditions,
-            rewrittenMetadata,
-            options.getContentType()) {
+            rewrittenMetadata) {
 
           @Override
           public Storage.Objects.Insert createRequest(InputStreamContent inputStream)
@@ -1120,10 +1122,12 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
    */
   private Storage.Objects.Insert prepareEmptyInsert(
       StorageResourceId resourceId, CreateObjectOptions createObjectOptions) throws IOException {
-    StorageObject object = new StorageObject();
-    object.setName(resourceId.getObjectName());
     Map<String, String> rewrittenMetadata = encodeMetadata(createObjectOptions.getMetadata());
-    object.setMetadata(rewrittenMetadata);
+    StorageObject object =
+        new StorageObject()
+            .setName(resourceId.getObjectName())
+            .setMetadata(rewrittenMetadata)
+            .setContentEncoding(createObjectOptions.getContentEncoding());
 
     // Ideally we'd use EmptyContent, but Storage requires an AbstractInputStreamContent and not
     // just an HttpContent, so we'll just use the next easiest thing.
@@ -2004,6 +2008,7 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
                         .setDestination(
                             new StorageObject()
                                 .setContentType(options.getContentType())
+                                .setContentEncoding(options.getContentEncoding())
                                 .setMetadata(encodeMetadata(options.getMetadata())))),
             destination.getBucketName());