+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageReadChannel.java
@@ -809,11 +809,11 @@ public class GoogleCloudStorageReadChannel implements SeekableByteChannel {
         !metadataInitialized,
         "can not initialize metadata, it already initialized for '%s'", resourceIdString);
     gzipEncoded = nullToEmpty(encoding).contains(GZIP_ENCODING);
-    if (gzipEncoded) {
-      size = Long.MAX_VALUE;
-    } else {
-      size = sizeFromMetadata;
+    if (gzipEncoded && !readOptions.getSupportGzipEncoding()) {
+      throw new IOException(
+          "Can't read GZIP encoded files - content encoding support is disabled.");
     }
+    size = gzipEncoded ? Long.MAX_VALUE : sizeFromMetadata;
     randomAccess = !gzipEncoded && readOptions.getFadvise() == Fadvise.RANDOM;
     checkEncodingAndAccess();