+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageTest.java
@@ -1195,11 +1195,11 @@ public class GoogleCloudStorageTest {
   }
 
   /**
-   * If we disable the supportContentEncoding option when opening a channel, and disable
-   * failing fast on nonexistent objects we should expect no extraneous metadata-GET calls at all.
+   * If we disable the supportGzipEncoding option when opening a channel, and disable failing fast
+   * on nonexistent objects we should expect no extraneous metadata-GET calls at all.
    */
   @Test
-  public void testOpenNoSupportContentEncodingAndNoFailFastOnNotFound() throws Exception {
+  public void testOpenNoSupportGzipEncodingAndNoFailFastOnNotFound() throws Exception {
     byte[] testData = { 0x01, 0x02, 0x03, 0x05, 0x08 };
 
     setUpBasicMockBehaviorForOpeningReadChannel(testData.length);